move tf header to EMTFTFHeader inside processing.py
